refactor: 기존 태그 추천 서비스 리네이밍
- 새로운 방식과 구분되도록
- 추상화. 어떻게가 아니라 무엇을 하는지 드러내도록.
